Move entrypoint.py to _internal
