+++ certbot-apache/certbot_apache/tests/configurator_test.py
@@ -121,12 +121,12 @@ class MultipleVhostsTest(util.ApacheTest):
         del os.environ["CERTBOT_DOCS"]
 
     def test_add_parser_arguments_all_configurators(self):  # pylint: disable=no-self-use
-        from certbot_apache.entrypoint import OVERRIDE_CLASSES
+        from certbot_apache._internal.entrypoint import OVERRIDE_CLASSES
         for cls in OVERRIDE_CLASSES.values():
             cls.add_parser_arguments(mock.MagicMock())
 
     def test_all_configurators_defaults_defined(self):
-        from certbot_apache.entrypoint import OVERRIDE_CLASSES
+        from certbot_apache._internal.entrypoint import OVERRIDE_CLASSES
         from certbot_apache.configurator import ApacheConfigurator
         parameters = set(ApacheConfigurator.OS_DEFAULTS.keys())
         for cls in OVERRIDE_CLASSES.values():
