+++ certbot-apache/certbot_apache/tests/util.py
@@ -14,7 +14,7 @@ from certbot.plugins import common
 from certbot.tests import util as test_util
 
 from certbot_apache import configurator
-from certbot_apache import entrypoint
+from certbot_apache._internal import entrypoint
 from certbot_apache._internal import obj
